allows files to be already deleted when removing tree

+++ dlt/common/storages/file_storage.py
@@ -345,6 +345,9 @@ class FileStorage:
 
     @staticmethod
     def rmtree_del_ro(action: AnyFun, name: str, exc: Any) -> Any:
+        # in rare cases file may be deleted between scandir and rm
+        if isinstance(exc[1], FileNotFoundError):
+            return
         if action in (os.unlink, os.remove, os.rmdir):
             # add +w with a mask
             os.chmod(name, os.stat(name).st_mode | stat.S_IWRITE)
